+++ Diva.Zengin.sln
@@ -6,7 +6,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Diva.Zengin.Tests", "Diva.Z
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Diva.Zengin.SourceGenerator", "Diva.Zengin.SourceGenerator\Diva.Zengin.SourceGenerator.csproj", "{90FBCCBF-1A47-4C60-8B76-7D739AC8DC2C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Diva.Zengin.Converters", "Diva.Zengin.Converters\Diva.Zengin.Converters.csproj", "{DFF9621A-FA81-4737-81FD-ADD1EF91F907}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Diva.Zengin.Helpers", "Diva.Zengin.Helpers\Diva.Zengin.Helpers.csproj", "{DFF9621A-FA81-4737-81FD-ADD1EF91F907}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
